Improve clarity between the different kinds of repositories used by the saga subsystem.
- Renamed `saga.Repository` to `InstanceRepository`
- Renamed `eventsourcing.Repository` to `InstanceRepository`
- Collapsed the snapshotting/non-snapshotting implementations of
  saga.InstanceRepository` into a single `StandardInstanceReposiotry` struct
