+++ src/ax/saga/instance.go
@@ -1,8 +1,11 @@
 package saga
 
 import (
+	"context"
+
 	"github.com/golang/protobuf/proto"
 	"github.com/jmalloc/ax/src/ax/ident"
+	"github.com/jmalloc/ax/src/ax/persistence"
 )
 
 // InstanceID uniquely identifies a saga instance.
@@ -41,4 +44,25 @@ type Data interface {
 	// Follow the same conventions as for error messages:
 	// https://github.com/golang/go/wiki/CodeReviewComments#error-strings
 	SagaDescription() string
+}
+
+// InstanceRepository is an interface for loading and saving saga instances.
+type InstanceRepository interface {
+	// LoadSagaInstance fetches a saga instance by its ID.
+	//
+	// If a saga instance is found; ok is true, otherwise it is false. A
+	// non-nil error indicates a problem with the store itself.
+	//
+	// It panics if the repository is not able to enlist in tx because it uses a
+	// different underlying storage system.
+	LoadSagaInstance(ctx context.Context, tx persistence.Tx, id InstanceID) (Instance, error)
+
+	// SaveSagaInstance persists a saga instance.
+	//
+	// It returns an error if the saga instance has been modified since it was
+	// loaded, or if there is a problem communicating with the store itself.
+	//
+	// It panics if the repository is not able to enlist in tx because it uses a
+	// different underlying storage system.
+	SaveSagaInstance(ctx context.Context, tx persistence.Tx, i Instance) error
 }
